Fix workflow-guard file-length budget: split liveness test fixtures, refresh budget
Split MobileShellRenderGridLivenessTests.swift (583 > 500 new-file
threshold) into the 4 regression tests (212 lines) plus
MobileShellRenderGridLivenessTestSupport.swift (381 lines: injected
clock, scripted host router, transport mocks, connected-store builder).

Budget refresh accepts the remaining growth as known debt:
MobileShellComposite.swift +238 (the watchdog probe machinery is
single-flight state coupled to the composite's private connection
state; extracting it would force ~15 members internal — precedent:
the composer merge refreshed this same file's budget) and
MobileHostService.swift +14 (alreadySubscribed ack field).
